+++ test/lib.test.js
@@ -4780,44 +4780,6 @@ describe('[order]', () => {
     );
   });
 
-  test('order.signRequestorder() preflightCheck default storage', async () => {
-    const signer = utils.getSignerFromPrivateKey(
-      tokenChainOpenethereumUrl,
-      getRandomWallet().privateKey,
-    );
-    const iexec = new IExec(
-      {
-        ethProvider: signer,
-      },
-      {
-        hubAddress,
-        resultProxyURL,
-        smsURL: smsMap,
-      },
-    );
-    const order = await iexec.order.createRequestorder({
-      app: getRandomAddress(),
-      category: 5,
-    });
-
-    await expect(iexec.order.signRequestorder(order)).rejects.toThrow(
-      Error(
-        'Requester storage token is not set for selected provider "ipfs". Result archive upload will fail.',
-      ),
-    );
-    await iexec.storage
-      .defaultStorageLogin()
-      .then(iexec.storage.pushStorageToken);
-    const res = await iexec.order.signRequestorder(order);
-    expect(res.salt).toMatch(bytes32Regex);
-    expect(res.sign).toMatch(signRegex);
-    expect(res).toEqual({
-      ...order,
-      ...{ params: JSON.stringify(order.params) },
-      ...{ sign: res.sign, salt: res.salt },
-    });
-  });
-
   test('order.signRequestorder() preflightCheck dropbox storage', async () => {
     const signer = utils.getSignerFromPrivateKey(
       tokenChainInstamineUrl,
@@ -6790,24 +6752,6 @@ describe('[order]', () => {
           tag: ['tee', 'scone'],
         },
       );
-
-      // trigger request check
-      await expect(
-        iexec.order.matchOrders({
-          apporder,
-          datasetorder,
-          workerpoolorder,
-          requestorder,
-        }),
-      ).rejects.toThrow(
-        Error(
-          'Requester storage token is not set for selected provider "ipfs". Result archive upload will fail.',
-        ),
-      );
-
-      await iexec.storage
-        .defaultStorageLogin()
-        .then(iexec.storage.pushStorageToken);
       const res = await iexec.order.matchOrders({
         apporder,
         datasetorder,
@@ -7013,14 +6957,6 @@ describe('[order]', () => {
         category: 1,
       })
       .then((o) => iexec.order.signRequestorder(o, { preflightCheck: false }));
-    await expect(iexec.order.publishRequestorder(requestorder)).rejects.toThrow(
-      Error(
-        'Requester storage token is not set for selected provider "ipfs". Result archive upload will fail.',
-      ),
-    );
-    await iexec.storage
-      .defaultStorageLogin()
-      .then(iexec.storage.pushStorageToken);
     const orderHash = await iexec.order.publishRequestorder(requestorder);
     expect(orderHash).toMatch(bytes32Regex);
   });